initial gtk4 port
Everything seems functional, sans dnd.

I don't fully understand what the color code does, but now it uses
Gdk.RGBA and there is now a new set_is_dark api which is expected
to be called by the app as you can't easily determine that yourself
anymore. If this would depend on libadwaita, that could be easily
replaced with it's style manager's is-dark property, which is exactly
for usecases like this.